+++ src/features/processes/stores/processes.store.ts
@@ -4,6 +4,7 @@ import { GlobalCpu, Process, streams, SysInfo } from "@/lib";
 import createSelectors from "@/utils/create-selectors";
 
 const DEFAULT_PROCESS: Process = {
+  children: [],
   cpuUsage: 0,
   memoryUsage: 0,
   name: "unknown",