+++ src/features/processes/stores/processes.store.ts
@@ -3,18 +3,29 @@ import { create } from "zustand";
 import { GlobalCpu, Process, streams, SysInfo } from "@/lib";
 import createSelectors from "@/utils/create-selectors";
 
-const DEFAULT_PROCESS: Process = {
+export const DEFAULT_PROCESS: Process = {
   cpuUsage: 0,
   memoryUsage: 0,
   name: "unknown",
-  pid: "1",
+  cmd: [],
+  exe: "unknown",
+  root: "unknown",
+  runTime: 0,
+  startTime: 0,
   status: "unknown",
+  diskUsage: {
+    readBytes: 0,
+    totalReadBytes: 0,
+    totalWrittenBytes: 0,
+    writtenBytes: 0,
+  },
+  timestamp: 0,
 };
 interface ProcessesState {
   processes: Process[];
   listen: () => void;
 }
-
+// TODO: Use enumerable store instead of this
 const useProcessesStore = create<ProcessesState>()((set, get) => ({
   processes: [DEFAULT_PROCESS],
   listen: () => streams.processes((processes) => set({ processes })),