+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## 0.9.5
-- Improve Theming options
+- Patch Network metrics not displaying
 
 ## 0.9.4
 - Fix CPU information not displaying