+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/).
 
+## 0.9.5
+- Patch Network metrics not displaying
+- Revamp Processes section
+- Created Processes comparitor
+- Create feature to do regular scan and turbo scan
+
 ## 0.9.4
-- Improve Theming options
 - Fix CPU information not displaying
 
 ## 0.9.3